archive.sh: improve logging (log to stderr), add long options, add option for compressor arguments.

+++ archive.sh
@@ -18,6 +18,10 @@ TARLESS=(
 	[7z]="1"
 )
 
+function log() {
+	echo "$*" >&2
+}
+
 function get_compressors() {
 	for compressor in "${!ARCHIVERS[@]}"; do
 		echo "$1$compressor: '${ARCHIVERS[$compressor]}'"
@@ -32,31 +36,34 @@ function usage() {
 	cat <<EOF
 archive.sh -- a directory archiving tool
 
-Usage: $0 [-fn] [-o OUTPUT] [-N NAME] [-c COMPRESSOR]
+Usage: $0 [-fn] [-o|--output OUTPUT] [-N|--name NAME] [-c|--compressor COMPRESSOR] [-O|--compressor-options OPTIONS]
 
 Options:
-	-o OUTPUT
+	-o, --output OUTPUT
 		Write archive to OUTPUT; may be directory or file.
 
-	-N NAME
+	-N, --name NAME
 		Use NAME for logging.
 		Also used as archive file name stem, if OUTPUT is a directory.
 
-	-c COMPRESSOR
+	-c, --compressor COMPRESSOR
 		Explicitly set compressing filter to COMPRESSOR.
 		Available options:
 $(get_compressors "		 * ")
 
-	-n
+	-n, --no-compressor
 		Explicitly disable compression.
 
-	-f
+	-f, --force
 		Allow overwriting the output file.
+
+	-O, --compressor-options OPTIONS
+		Options for the compressor.
 
 EOF
 }
 
-ARGS=$(getopt -- "o:N:nc:f" "$@")
+ARGS=$(getopt --long "output:,name:,compressor:,no-compressor,force,compressor-options" -- "o:N:nc:fO:" "$@")
 if (( "$?" )); then
 	exit 1
 fi
@@ -66,24 +73,30 @@ eval set -- "$ARGS"
 # Defaults.
 
 ARCHIVER=""
-PV_OPTS="-pba"
-PV_SECONDSTAGE_OPTS="-b"
+PV_OPTS="-pba -W"
+PV_SECONDSTAGE_OPTS="-ba -W"
+ARCHIVER_OPTS=""
 
 while true; do
 	case "$1" in
-		-f)
+		-f|--force)
 			FORCE=1
 			;;
-		-o)
+		-o|--output)
 			shift
+			if [[ ! "$OUTPUT" ]]; then
+				log "E: empty output name."
+				exit 1
+			fi
+
 			OUTPUT="$1"
 
-			if [[ ! -d "$OUTPUT" ]]; then
+			if [[ "$OUTPUT" != "-" && ! -d "$OUTPUT" ]]; then
 				if [[ -z "$NAME" ]]; then
 					NAME="${OUTPUT##*/}" # use output as name: remove directories
 					NAME="${NAME%%.*}" # use output as name: remove extensions
 
-					echo "N: defaulting name to '$NAME' from output file stem."
+					log "N: defaulting name to '$NAME' from output file stem."
 				fi
 
 				if [[ "$OUTPUT" =~ \.([^.]*)$ ]]; then
@@ -92,33 +105,37 @@ while true; do
 					if [[ -z "$ARCHIVER" && -n "${ARCHIVERS[$EXTENSION]}" ]]; then
 						ARCHIVER="$EXTENSION"
 
-						echo "N: defaulting archiver to '$ARCHIVER' from output file extension."
+						log "N: defaulting archiver to '$ARCHIVER' from output file extension."
 					fi
 				fi
 			fi
 			;;
-		-N)
+		-N|--name)
 			shift
 			NAME="$1"
 			;;
-		-n)
+		-n|--no-compressor)
 			ARCHIVER=""
 			;;
-		-c)
+		-c|--compressor)
 			shift
 			if [[ "${ARCHIVERS[$1]}" ]]; then
 				ARCHIVER="$1"
 			else
-				echo "E: invalid archiver '$1'."
+				log "E: invalid archiver '$1'."
 				exit 1
 			fi
+			;;
+		-O|--compressor-options)
+			shift
+			ARCHIVER_OPTS="$1"
 			;;
 		--)
 			shift # we won't get to the shift in the end of loop
 			break
 			;;
 		*)
-			echo "E: wrong option."
+			log "E: wrong option."
 			usage
 			exit 1
 			;;
@@ -128,15 +145,21 @@ while true; do
 done
 
 if ! (( $# )); then
-	echo "E: No source files to archive."
+	log "E: No source files to archive."
 	usage
 	exit 1
 fi
 
 if [[ -z "$NAME" ]]; then
-	NAME="$(basename "$1")" # use first source directory as name: remove directories
+	if (( $# == 1 )); then
+		NAME="$(basename "$1")" # use first input name: remove directories
 
-	echo "N: defaulting name to '$NAME' from first source directory."
+		log "N: defaulting name to '$NAME' from the only input."
+	else
+		NAME="files"
+
+		log "N: defaulting name to '$NAME'."
+	fi
 fi
 
 if [[ "$ARCHIVER" ]]; then
@@ -149,45 +172,76 @@ else
 	EXTENSION=".tar"
 fi
 
-if [[ -z "$OUTPUT" ]]; then
-	OUTPUT="${NAME}${EXTENSION}"
+if [[ "$OUTPUT" != "-" ]]; then
+	if [[ -z "$OUTPUT" ]]; then
+		OUTPUT="${NAME}${EXTENSION}"
 
-	echo "N: defaulting output to '$OUTPUT'."
-elif [[ -d "$OUTPUT" ]]; then
-	OUTPUT+="/${NAME}${EXTENSION}"
+		log "N: defaulting output to '$OUTPUT'."
+	elif [[ -d "$OUTPUT" ]]; then
+		OUTPUT+="/${NAME}${EXTENSION}"
 
-	echo "N: defaulting output to '$OUTPUT'."
-fi
+		log "N: defaulting output to '$OUTPUT'."
+	fi
 
-if [[ -e "$OUTPUT" ]]; then
-	if (( FORCE )); then
-		echo "W: '$OUTPUT' exists and is not a directory. Overwriting."
-		rm -f "$OUTPUT"
-	else
-		echo "E: '$OUTPUT' exists and is not a directory. Pass '-f' to overwrite."
+	if [[ -d "$OUTPUT" ]]; then
+		log "E: '$OUTPUT' exists and is a directory. Aborting."
 		exit 1
+	elif [[ -e "$OUTPUT" ]]; then
+		if (( FORCE )); then
+			if [[ -f "$OUTPUT" ]]; then
+				log "W: '$OUTPUT' exists and is a regular file. Removing."
+				rm -f "$OUTPUT"
+			else
+				log "W: '$OUTPUT' exists and is not a regular file. The archiver will decide what to do."
+			fi
+		else
+			log "E: '$OUTPUT' exists. Pass '-f' to overwrite."
+			exit 1
+		fi
 	fi
 fi
+
+function tar_and_archive_cmd() {
+	local size="$(get_size "$@")"
+	tar -c "$@" \
+		| pv ${PV_OPTS}             -cN "Archiving   $NAME (tar)" -s "$size" \
+		| $ARCHIVER_CMD $ARCHIVER_OPTS \
+		| pv ${PV_SECONDSTAGE_OPTS} -WN "Compressing $NAME ($ARCHIVER_NAME)"
+}
+
+function tar_cmd() {
+	local size="$(get_size "$@")"
+	tar -c "$@" \
+		| pv $PV_OPTS -N "$NAME" -s "$size"
+}
 
 if [[ "$ARCHIVER" ]]; then
 	ARCHIVER_CMD="${ARCHIVERS["$ARCHIVER"]}"
 	ARCHIVER_NAME="${ARCHIVER_CMD%% *}" # Select first word of the string (command name)
 
-	echo "N: compressing and writing to '$OUTPUT'"
-
 	if [[ "${TARLESS[$ARCHIVER]}" ]]; then
-		$ARCHIVER_CMD "${OUTPUT}" "$@"
+		if [[ "$OUTPUT" == "-" ]]; then
+			log "E: writing to stdout is not supported with this archiver."
+			exit 1
+		else
+			log "N: compressing and writing to '$OUTPUT'"
+			$ARCHIVER_CMD $ARCHIVER_OPTS "${OUTPUT}" "$@"
+		fi
 	else
-		tar -c "$@" \
-			| pv ${PV_OPTS}             -cN "Archiving   $NAME (tar)" -s $(get_size "$@") \
-			| $ARCHIVER_CMD \
-			| pv ${PV_SECONDSTAGE_OPTS} -WN "Compressing $NAME ($ARCHIVER_NAME)" \
-			> "${OUTPUT}"
+		if [[ "$OUTPUT" == "-" ]]; then
+			log "N: compressing and writing to stdout"
+			tar_and_archive_cmd "$@"
+		else
+			log "N: compressing and writing to '$OUTPUT'"
+			tar_and_archive_cmd "$@" > "$OUTPUT"
+		fi
 	fi
 else
-	echo "N: writing to '${OUTPUT}'"
-
-	tar -c "$@" \
-		| pv $PV_OPTS -N "$NAME" -s $(get_size "$@") \
-		> "${OUTPUT}"
+	if [[ "$OUTPUT" == "-" ]]; then
+		log "N: writing to stdout"
+		tar_cmd "$@"
+	else
+		log "N: writing to '$OUTPUT'"
+		tar_cmd "$@" > "$OUTPUT"
+	fi
 fi
